introduce a new error reporting protocol and sentry error reporter class to handle sending of errors to sentry dashboard.
